src/gap_scanner/pyproject.toml
[project]
name = "gap-scanner"
version = "0.1.0"
description = "gap_scanner"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "tinkoff-investments>=0.2.0b115",
]
